+++ Production/govuk_ios/Validators/MaximumLengthValidator.swift
@@ -1,0 +1,14 @@
+import Foundation
+
+struct MaximumLengthValidator: ValidatorProvider {
+    private let length: Int
+
+    init(length: Int) {
+        self.length = length
+    }
+
+    func validate(input: String?) -> Bool {
+        guard let input = input else { return false }
+        return input.count <= length
+    }
+}
