+++ Production/govuk_ios/Validators/MinimumLengthValidator.swift
@@ -1,0 +1,14 @@
+import Foundation
+
+struct MinimumLengthValidator: ValidatorProvider {
+    private let requiredStringLength: Int
+
+    init(requiredStringLength: Int) {
+        self.requiredStringLength = requiredStringLength
+    }
+
+    func validate(input: String?) -> Bool {
+        guard let input = input else { return false }
+        return input.count >= requiredStringLength
+    }
+}
