+++ Production/govuk_ios/Validators/ValidatorProvider.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+protocol ValidatorProvider {
+    func validate(input: String?) -> Bool
+}
